Patch to guard against invalid event state in Firestore
If an event participant isn't present in our users collection it leads to page rendering errors on the eventPage and the corresponding user's profile page, inserted checks to guard against this. Stretch extension: would be nice to have a backend service that checks our collections for invalid state occurences and updates the state as needed.